Splitting IRC tests from core Connect tests, for clarity
